use supersession to mock session
